# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty 100 Analytics API",
    "description": "REST API for the Nifty 100 fundamental analysis project.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "description": "Basic liveness and status check for the API.",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Companies",
        "description": "List all companies, with optional filters on sector, market cap category or a name/ticker search.",
        "operationId": "list_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          },
          {
            "name": "market_cap_category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Market Cap Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Profile",
        "description": "Full company profile: company info + latest year KPIs + sector data.",
        "operationId": "company_profile_api_v1_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/pl": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Profit And Loss",
        "description": "Profit and loss history. from_year/to_year are in YYYY-MM format.",
        "operationId": "profit_and_loss_api_v1_companies__ticker__pl_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "From Year"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "To Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/bs": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Balance Sheet",
        "description": "Balance sheet history. from_year/to_year are in YYYY-MM format.",
        "operationId": "balance_sheet_api_v1_companies__ticker__bs_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "From Year"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "To Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/cashflow": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Cash Flow",
        "description": "Cash flow history. from_year/to_year are in YYYY-MM format.",
        "operationId": "cash_flow_api_v1_companies__ticker__cashflow_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "From Year"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "To Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/ratios": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Ratios",
        "description": "All computed KPIs per year for the company, or a single year if `year` is given.",
        "operationId": "ratios_api_v1_companies__ticker__ratios_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/tearsheet": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Tearsheet",
        "description": "Download the pre-generated 2-page tearsheet PDF for this company.",
        "operationId": "tearsheet_api_v1_companies__ticker__tearsheet_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Peer Compare",
        "description": "Radar data: this company's 8 axis metrics + the peer group average + the benchmark company.",
        "operationId": "peer_compare_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Documents",
        "description": "Annual report links for this company, each flagged with whether the URL still works.",
        "operationId": "documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener": {
      "get": {
        "tags": [
          "screener"
        ],
        "summary": "Screener",
        "description": "Filter the 92-company universe. All parameters are optional and combine with AND.\nAny non-numeric value for a numeric parameter returns HTTP 400.",
        "operationId": "screener_api_v1_screener_get",
        "parameters": [
          {
            "name": "min_roe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Min Roe"
            }
          },
          {
            "name": "max_de",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Max De"
            }
          },
          {
            "name": "min_fcf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Min Fcf"
            }
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          },
          {
            "name": "min_rev_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Min Rev Cagr 5Yr"
            }
          },
          {
            "name": "min_pat_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Min Pat Cagr 5Yr"
            }
          },
          {
            "name": "max_pe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Max Pe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sectors": {
      "get": {
        "tags": [
          "sectors"
        ],
        "summary": "List Sectors",
        "description": "All broad sectors with company count and median ROE / P/E / D/E.",
        "operationId": "list_sectors_api_v1_sectors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "tags": [
          "sectors"
        ],
        "summary": "Sector Companies",
        "description": "All companies in a sector with their latest year KPIs.",
        "operationId": "sector_companies_api_v1_sectors__sector__companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{group_name}": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "Peer Group",
        "description": "All companies in a peer group with their percentile rank for each of the 10 metrics.",
        "operationId": "peer_group_api_v1_peers__group_name__get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-cap/{ticker}": {
      "get": {
        "tags": [
          "valuation"
        ],
        "summary": "Market Cap History",
        "description": "Historical valuation multiples (P/E, P/B, EV/EBITDA, dividend yield) from 2019 to 2024.",
        "operationId": "market_cap_history_api_v1_market_cap__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/stats": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Portfolio Stats",
        "description": "P10 through P90 percentile table for the 10 core KPIs across all 92 companies.",
        "operationId": "portfolio_stats_api_v1_portfolio_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}